requirements.txt
Django==4.1.7
google_api_python_client==2.80.0
google_auth_oauthlib==1.0.0
protobuf==4.22.0
